Fix GUI using wrong menu index to update controls
Due to a misinterpretation of the list selection changed event's index
properties the wrong row of the list was being used to update the item
controls.

The indexes of the event are the min and max of the changes from
previously selected and newly selected (i.e. not convenient to
consistently obtain the newly selected item).

This way gets the selected index from the model which is consistent.