Safeguard againt concurrent modification of lists
There was some flawed logic in our previous solution for this (only in
the server methods previously) because the methods would need to be
synchronized for it to work consistently (if two observers were being
registered, one could be theoretically missed if the thread paused after
copying the list and another wrote the list before it resumed).

This way the observer list changes and iteration is synchronized using
the list as the lock (most significant for the queue since we don't want
to lock the whole object).

The queue's getLane method also now returns a copy of the lane
collection to avoid the same error if the calling code later wants to
iterate it (which ours do).

Fixes #110